{"schema":"decodex/commit/1","summary":"Split Decodex app bridge modules","authority":"manual"}
